requirements.txt
aiogram>=3.0.0
openai>=1.0.0
httpx>=0.25.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
watchdog>=4.0.0
